Move EventId in scenario.
