+++ sdk/compiler/scenario-service/server/src/main/scala/com/digitalasset/daml/lf/scenario/Conversions.scala
@@ -6,7 +6,6 @@ package scenario
 
 import com.digitalasset.daml.lf.data.{ImmArray, Numeric, Ref}
 import com.digitalasset.daml.lf.language.Ast.PackageMetadata
-import com.digitalasset.daml.lf.ledger.EventId
 import com.digitalasset.daml.lf.scenario.api.{v1 => proto}
 import com.digitalasset.daml.lf.speedy.{SError, SValue, TraceLog, Warning, WarningLog}
 import com.digitalasset.daml.lf.transaction.{
