+++ sdk/daml-lf/scenario-interpreter/src/main/scala/com/digitalasset/daml/lf/scenario/Error.scala
@@ -7,7 +7,6 @@ package scenario
 import com.digitalasset.daml.lf.data.Ref.{Identifier, Party, PackageId}
 import com.digitalasset.daml.lf.data.Time
 import com.digitalasset.daml.lf.language.Ast.PackageMetadata
-import com.digitalasset.daml.lf.ledger.EventId
 import com.digitalasset.daml.lf.speedy.SError.SError
 import com.digitalasset.daml.lf.transaction.{GlobalKey, VersionedTransaction}
 import com.digitalasset.daml.lf.value.Value.ContractId
